4.1.6 Вынесет функции общего назначения из util в common
